change base image to elixir and then install ROS
Since `erlang-solutions_2.0_all.deb` package, that has been employed for this repository, sometimes exclude the versions of Elixir required for testing Rclex. Since we are planning to automate Docker build & push in the future, this mannar will be a cause to achieve this. Therefore, we change the building policy:
- before: employ `ros` from OSRF as base image, and then install Elixir/Erlang environment
- after : employ `elixir` from Hexpm team as base image, and then install ROS environment

This change also contribute to reduce the size of image, from 1.63 GB to 1.02 GB (in `foxy-ex1.12.3-otp24.1.5`).